requirements.txt
asgiref==3.3.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.3
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.3.0
djoser==2.1.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
psycopg2==2.8.6
drf_yasg==1.20.0
inflection==0.5.1
pyparsing==2.4.7
ruamel.yaml==0.16.12
django-cors-headers==3.5.0
django-js-asset==1.2.2
django-mptt==0.11.0